move ApiClientHelpers to its own module
